requirements.txt
httpx>=0.27.0
python-dotenv>=1.0.1
pytest>=8.0.0
pytest-cov>=5.0.0
pre-commit>=3.0.0 